refactor circle button to own file
